Fixed fftw.h -> sfftw.h, double -> float, etc., following PME code.

+++ src/SimParameters.C
@@ -23,8 +23,8 @@
 #include "InfoStream.h"
 #include <time.h>
 #ifdef NAMD_FFTW
-#include <fftw.h>
-#include <rfftw.h>
+#include <sfftw.h>
+#include <srfftw.h>
 #endif
 #if defined(WIN32) && !defined(__CYGWIN__)
 #include <direct.h>
@@ -3050,8 +3050,8 @@ void SimParameters::print_config(ParseOptions &opts, ConfigList *config, char *&
 
        int n[3]; n[0] = PMEGridSizeX; n[1] = PMEGridSizeY; n[2] = PMEGridSizeZ;
        fftw_complex *work = new fftw_complex[n[0]];
-       double *grid1 = new double[n[1]*n[2]];
-       double *grid2 = new double[n[0]*block2*dim3];
+       float *grid1 = new float[n[1]*n[2]];
+       float *grid2 = new float[n[0]*block2*dim3];
        iout << iINFO << "Optimizing 6 FFT steps.  1..." << endi;
        rfftwnd_destroy_plan( rfftwnd_create_plan_specific(
 	 2, n+1, FFTW_REAL_TO_COMPLEX,
